+++ node.h
@@ -8,7 +8,7 @@
 class Node
 {
 public:
-	Node(Point* _point, std::shared_ptr<Node> _parent = NULL, int _fCost = 0, int _distanceWithStart = 0)
+	Node(Point* _point, std::shared_ptr<Node> _parent = NULL, double _fCost = 0, double _distanceWithStart = 0)
 	{
 		point = _point;
 		parent = _parent;
@@ -17,17 +17,17 @@ public:
 	}
 	std::shared_ptr<Node> getParent() const;
 	Point* getPoint() const;
-	int getFCost() const;
-	int getDistanceWithStart() const;
+	double getFCost() const;
+	double getDistanceWithStart() const;
 	void setParent(std::shared_ptr<Node> _parent);
 	void setPoint(Point* _point);
-	void setFCost(int _fCost);
-	void setDistanceWithStart(int _distanceWithStart);
+	void setFCost(double _fCost);
+	void setDistanceWithStart(double _distanceWithStart);
 
 
 private:
 	std::shared_ptr<Node> parent;
 	Point* point;
-	int fCost;
-	int distanceWithStart;
+	double fCost;
+	double distanceWithStart;
 };
